requirements.txt
django 
djangorestframework
django-environ
django-markdown
django-filter
drf_social_oauth2